@@ -4,16 +4,15 @@ version: 1.2.4
 homepage: https://github.com/RafaelBarbosatec/tutorial_coach_mark
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  dart_code_metrics: ^5.0.0
+  dfc_lints:
+    git: https://github.com/sgehrman/dfc_lints
 
 flutter:
-
-
