@@ -8,7 +8,6 @@ import 'package:tutorial_coach_mark/src/widgets/animated_focus_light.dart';
 
 class TutorialCoachMarkWidget extends StatefulWidget {
   const TutorialCoachMarkWidget({
-    Key? key,
     required this.targets,
     this.finish,
     this.paddingFocus = 10,
@@ -16,7 +15,7 @@ class TutorialCoachMarkWidget extends StatefulWidget {
     this.onClickTargetWithTapPosition,
     this.clickOverlay,
     this.alignSkip = Alignment.bottomRight,
-    this.textSkip = "SKIP",
+    this.textSkip = 'SKIP',
     this.onClickSkip,
     this.colorShadow = Colors.black,
     this.opacityShadow = 0.8,
@@ -28,13 +27,15 @@ class TutorialCoachMarkWidget extends StatefulWidget {
     this.pulseVariation,
     this.pulseEnable = true,
     this.skipWidget,
-  })  : assert(targets.length > 0),
+    Key? key,
+  })  : assert(targets.length > 0, 'no targets'),
         super(key: key);
 
   final List<TargetFocus> targets;
-  final FutureOr Function(TargetFocus)? clickTarget;
-  final FutureOr Function(TargetFocus, TapDownDetails)? onClickTargetWithTapPosition;
-  final FutureOr Function(TargetFocus)? clickOverlay;
+  final FutureOr<void> Function(TargetFocus)? clickTarget;
+  final FutureOr<void> Function(TargetFocus, TapDownDetails)?
+      onClickTargetWithTapPosition;
+  final FutureOr<void> Function(TargetFocus)? clickOverlay;
   final Function()? finish;
   final Color colorShadow;
   final double opacityShadow;
@@ -55,7 +56,8 @@ class TutorialCoachMarkWidget extends StatefulWidget {
   TutorialCoachMarkWidgetState createState() => TutorialCoachMarkWidgetState();
 }
 
-class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implements TutorialCoachMarkController {
+class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget>
+    implements TutorialCoachMarkController {
   final GlobalKey<AnimatedFocusLightState> _focusLightKey = GlobalKey();
   bool showContent = false;
   TargetFocus? currentTarget;
@@ -82,7 +84,8 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
               return widget.clickTarget?.call(target);
             },
             clickTargetWithTapPosition: (target, tapDetails) {
-              return widget.onClickTargetWithTapPosition?.call(target, tapDetails);
+              return widget.onClickTargetWithTapPosition
+                  ?.call(target, tapDetails);
             },
             clickOverlay: (target) {
               return widget.clickOverlay?.call(target);
@@ -101,10 +104,10 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
           ),
           AnimatedOpacity(
             opacity: showContent ? 1 : 0,
-            duration: Duration(milliseconds: 300),
+            duration: const Duration(milliseconds: 300),
             child: _buildContents(),
           ),
-          _buildSkip()
+          _buildSkip(),
         ],
       ),
     );
@@ -112,17 +115,17 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
 
   Widget _buildContents() {
     if (currentTarget == null) {
-      return SizedBox.shrink();
+      return const SizedBox.shrink();
     }
 
     List<Widget> children = <Widget>[];
 
     final target = getTargetCurrent(currentTarget!);
     if (target == null) {
-      return SizedBox.shrink();
+      return const SizedBox.shrink();
     }
 
-    var positioned = Offset(
+    final positioned = Offset(
       target.offset.dx + target.size.width / 2,
       target.offset.dy + target.size.height / 2,
     );
@@ -130,8 +133,10 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
     double haloWidth;
     double haloHeight;
 
-    if (currentTarget!.shape == ShapeLightFocus.Circle) {
-      haloWidth = target.size.width > target.size.height ? target.size.width : target.size.height;
+    if (currentTarget!.shape == ShapeLightFocus.circle) {
+      haloWidth = target.size.width > target.size.height
+          ? target.size.width
+          : target.size.height;
       haloHeight = haloWidth;
     } else {
       haloWidth = target.size.width;
@@ -141,7 +146,7 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
     haloWidth = haloWidth * 0.6 + widget.paddingFocus;
     haloHeight = haloHeight * 0.6 + widget.paddingFocus;
 
-    double weight = 0.0;
+    double weight = 0;
     double? top;
     double? bottom;
     double? left;
@@ -162,7 +167,8 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
             weight = MediaQuery.of(context).size.width;
             left = 0;
             top = null;
-            bottom = haloHeight + (MediaQuery.of(context).size.height - positioned.dy);
+            bottom = haloHeight +
+                (MediaQuery.of(context).size.height - positioned.dy);
           }
           break;
         case ContentAlign.left:
@@ -197,11 +203,13 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
         bottom: bottom,
         left: left,
         right: right,
-        child: Container(
+        child: SizedBox(
           width: weight,
           child: Padding(
             padding: i.padding,
-            child: i.builder != null ? i.builder?.call(context, this) : (i.child ?? SizedBox.shrink()),
+            child: i.builder != null
+                ? i.builder?.call(context, this)
+                : (i.child ?? const SizedBox.shrink()),
           ),
         ),
       );
@@ -214,18 +222,19 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
 
   Widget _buildSkip() {
     if (widget.hideSkip!) {
-      return SizedBox.shrink();
+      return const SizedBox.shrink();
     }
+
     return Align(
       alignment: currentTarget?.alignSkip ?? widget.alignSkip,
       child: SafeArea(
         child: AnimatedOpacity(
           opacity: showContent ? 1 : 0,
-          duration: Duration(milliseconds: 300),
+          duration: const Duration(milliseconds: 300),
           child: InkWell(
             onTap: skip,
             child: Padding(
-              padding: const EdgeInsets.all(20.0),
+              padding: const EdgeInsets.all(20),
               child: IgnorePointer(
                 child: widget.skipWidget ??
                     Text(
@@ -240,9 +249,12 @@ class TutorialCoachMarkWidgetState extends State<TutorialCoachMarkWidget> implem
     );
   }
 
+  @override
   void skip() => widget.onClickSkip?.call();
 
+  @override
   void next() => _focusLightKey.currentState?.next();
 
+  @override
   void previous() => _focusLightKey.currentState?.previous();
 }
